requirements.txt
datasets
chromadb
sentence-transformers
openai
pyyaml
tqdm
requests
numpy
accelerate